move redis cache
