+++ backend/domain/model/top_page.go
@@ -1,0 +1,1 @@
+package model
